+++ src/rest/rest_handlers.rs
@@ -80,7 +80,7 @@ pub async fn get_persons(aggregator: MutexAggregator) -> Result<Box<dyn Reply>, 
     }
 }
 
-pub async fn get_person_events(aggregator: MutexAggregator, repeat_every_secs: u64, from_revision: Option<u32>) -> Result<impl Reply, Infallible> {
+pub async fn get_person_events(aggregator: MutexAggregator, repeat_every_secs: u64, from_revision: Option<usize>) -> Result<impl Reply, Infallible> {
     let from_revision = from_revision.unwrap_or(1);
     let fetcher = Box::new(PersonEventFetcher::new(aggregator, from_revision));
     let stream = ScheduledStream::new(Duration::from_secs(repeat_every_secs), fetcher);
@@ -90,7 +90,7 @@ pub async fn get_person_events(aggregator: MutexAggregator, repeat_every_secs: u
     Ok(sse::reply(stream))
 }
 
-pub async fn get_location_events(aggregator: MutexAggregator, repeat_every_secs: u64, from_revision: Option<u32>) -> Result<impl Reply, Infallible> {
+pub async fn get_location_events(aggregator: MutexAggregator, repeat_every_secs: u64, from_revision: Option<usize>) -> Result<impl Reply, Infallible> {
     let from_revision = from_revision.unwrap_or(1);
     let fetcher = Box::new(LocationEventFetcher::new(aggregator, from_revision));
     let stream = ScheduledStream::new(Duration::from_secs(repeat_every_secs), fetcher);
